+++ src/io/css/hsl2css.js
@@ -12,16 +12,16 @@ const rnd = (a) => Math.round(a * 100) / 100;
 const hsl2css = (...args) => {
     const hsla = unpack(args, 'hsla');
     let mode = last(args) || 'lsa';
-    hsla[0] = rnd(hsla[0] || 0);
+    hsla[0] = rnd(hsla[0] || 0) + 'deg';
     hsla[1] = rnd(hsla[1] * 100) + '%';
     hsla[2] = rnd(hsla[2] * 100) + '%';
     if (mode === 'hsla' || (hsla.length > 3 && hsla[3] < 1)) {
-        hsla[3] = hsla.length > 3 ? hsla[3] : 1;
+        hsla[3] = '/ ' + (hsla.length > 3 ? hsla[3] : 1);
         mode = 'hsla';
     } else {
         hsla.length = 3;
     }
-    return `${mode}(${hsla.join(',')})`;
+    return `${mode.substr(0, 3)}(${hsla.join(' ')})`;
 };
 
 export default hsl2css;