+++ src/io/lab/index.js
@@ -4,12 +4,15 @@ import Color from '../../Color.js';
 import input from '../input.js';
 import lab2rgb from './lab2rgb.js';
 import rgb2lab from './rgb2lab.js';
+import { getLabWhitePoint, setLabWhitePoint } from './lab-constants.js';
 
 Color.prototype.lab = function () {
     return rgb2lab(this._rgb);
 };
 
 chroma.lab = (...args) => new Color(...args, 'lab');
+chroma.setLabWhitePoint = setLabWhitePoint;
+chroma.getLabWhitePoint = getLabWhitePoint;
 
 input.format.lab = lab2rgb;